% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_collapse.R
\name{spq_collapse}
\alias{spq_collapse}
\title{Title (TBD)}
\usage{
spq_collapse(x, window, fun = c, ...)
}
\arguments{
\item{x}{An object of class SpatRaster (terra)}

\item{window}{A length of the side of a square-shaped block of cells.
Expressed in the numbers of cells, it defines the extent of a local pattern.}

\item{fun}{Function to summarize the values in a given \code{window}.
The default, \code{c}, returns all of the values inside specified windows.}

\item{...}{Additional arguments for \code{terra::extract} or for the specified function}
}
\value{
An object of class SpatRaster (terra)
}
\description{
Title (TBD)
}
\examples{
library(terra)
x = rast(system.file("raster/ta_scaled.tif", package = "spquery"))
xc = spq_collapse(x, window = 5)
plot(xc)
}
