% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_compare.R
\name{spq_compare}
\alias{spq_compare}
\title{Compares values between two rasters}
\usage{
spq_compare(x, y, dist_fun, ...)
}
\arguments{
\item{x}{An object of class SpatRaster (terra)}

\item{y}{An object of class SpatRaster (terra)}

\item{dist_fun}{Distance measure used. This function uses the \code{philentropy::distance} function (run \code{philentropy::getDistMethods()} to find possible distance measures), \code{proxy::dist}  in the background.
It is also possible to use \code{"dtw"} (dynamic time warping)}

\item{...}{Additional arguments for \code{philentropy::dist_one_one}, \code{proxy::dist}, or \code{dtwclust::dtw_basic}.
When \code{dist_fun = "dtw"} is used, \code{ndim} should be set to specify how many dimension the input raster time-series has.}
}
\value{
An object of class SpatRaster (terra)
}
\description{
Compares values between two rasters based on a given distance measure.
}
\examples{
library(terra)
library(sf)
ta = rast(system.file("raster/ta_scaled.tif", package = "spquery"))
pr = rast(system.file("raster/pr_scaled.tif", package = "spquery"))

re = spq_compare(ta, pr, dist_fun = "jensen-shannon")
plot(re)

}
