% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_search.R
\name{spq_search}
\alias{spq_search}
\title{Search for areas with similar values}
\usage{
spq_search(x, y, dist_fun, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{An object of class SpatRaster (terra)}

\item{dist_fun}{Distance measure used. This function uses the \code{philentropy::distance} function (run \code{philentropy::getDistMethods()} to find possible distance measures), \code{proxy::dist}  in the background.
It is also possible to use \code{"dtw"} (dynamic time warping)}

\item{...}{Additional arguments for \code{philentropy::dist_one_one}, \code{proxy::dist}, or \code{dtwclust::dtw_basic}.
When \code{dist_fun = "dtw"} is used, \code{ndim} should be set to specify how many dimension the input raster time-series has.}
}
\value{
An object of class SpatRaster (terra)
}
\description{
Searches for areas with similar values to a query.
}
\examples{
library(terra)
library(sf)
ta = rast(system.file("raster/ta_scaled.tif", package = "spquery"))
pr = rast(system.file("raster/pr_scaled.tif", package = "spquery"))
twor = c(ta, pr)
london = st_sf(geom = st_sfc(st_point(c(-0.1, 51.5))), crs = "EPSG:4326")
london_vector = as.numeric(extract(twor, london))

re = spq_search(london_vector, twor, dist_fun = "euclidean")
plot(re)

#re2 = spq_search(london_vector, twor, dist_fun = "dtw", ndim = 2, norm = "L2")
#plot(re2)
}
