% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spquery-package.R
\docType{package}
\name{spquery-package}
\alias{spquery}
\alias{spquery-package}
\title{spquery: Queries on Spatial Raster Data}
\description{
Performs several queries based on spatial raster data. It allows searching for areas with similar values to a query and comparing values between two rasters based on a given distance measure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jakubnowosad.com/spquery/}
  \item Report bugs at \url{https://github.com/Nowosad/spquery/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
