% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq_search.R
\name{spq_search}
\alias{spq_search}
\title{Title}
\usage{
spq_search(x, y, dist_fun)
}
\arguments{
\item{dist_fun}{}
}
\description{
Title
}
\examples{
library(terra)
library(sf)
ta = rast(system.file("raster/ta_scaled.tif", package = "spquery"))
pr = rast(system.file("raster/pr_scaled.tif", package = "spquery"))
twor = c(ta, pr)
london = st_sf(geom = st_sfc(st_point(c(-0.1, 51.5))), crs = "EPSG:4326")
london_rast = crop(twor, vect(london))
london_vector = as.numeric(values(london_rast))

re = spq_search(twor, london_vector, dist_fun = "euclidean")
plot(re)

#rd = spq_search(twor, london_vector, dist_fun = "dtw")
#plot(rd)

#rd2 = spq_search(twor, matrix(london_vector, ncol = 2), dist_fun = "dtw")
#plot(rd2)
}
